requirements.txt
python-telegram-bot==20.7
requests==2.31.0
PyJWT==2.8.0
aiohttp==3.9.1
cryptography==42.0.5